pyproject.toml
[tool.poetry]
name = "deb-distro"
version = "0.0.1"
description = "a framework used to create custom debian distributions"
authors = [""]
license = "Apache-2.0"

[tool.poetry.dependencies]
python = "^3.9.6"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
black = "^22.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
